+++ R/boundary_matrix.R
@@ -240,7 +240,7 @@ boundary_matrix.sf <- function(x, ...) {
 boundary_matrix.default <- function(x, ...) {
   cli::cli_abort(
     c(
-      "!" = "{.arg x} must be a {.cls sf} or {.cls SpatRaster}.",
+      "{.arg x} must be a {.cls sf} or {.cls SpatRaster}.",
       "x" = "{.arg x} is a {.cls {class(x)}}."
     )
   )